etcd-tester: use Hash method to get both revision and hash
